requirements.txt
APScheduler==3.10.0
asgiref==3.5.0
Django==4.0.2
django-js-asset==2.0.0
django-mptt==0.14.0
django-numerators==0.0.2
django-utils-six==2.0
django-widget-tweaks==1.4.12
Pillow==9.4.0
pyotp==2.8.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.2
whitenoise==6.0.0
